Cloud Function Migration/requirements.txt

google-cloud-pubsub
google-cloud-storage
journalist

